접수 승인요청 기능 구현

+++ src/main/java/lims/api/test/entity/Receipt.java
@@ -14,7 +14,7 @@ public class Receipt {
     private String receiptName;
     private String receiptNumber;
     private List<TestItem> testItems = new ArrayList<>();
-    private Integer approveId;
+    private Long approveId;
 
     public boolean isNew() {
         return this.id == null || this.id == 0;
